pyproject.toml
# Why pyproject.toml?
# https://stackoverflow.com/a/67103294
# https://ianhopkinson.org.uk/2022/02/understanding-setup-py-setup-cfg-and-pyproject-toml-in-python/

# https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html
# https://packaging.python.org/en/latest/specifications/declaring-project-metadata/
[build-system]
requires = ["setuptools >= 62.6"]
build-backend = "setuptools.build_meta"

[project]
name = "cd2root"
description = "Change Working Directory to the Project Root"
requires-python = ">=3.8"
license = { text = "MIT license" }
authors = [{ name = "garywei944", email = "garywei944@gmail.com" }]
keywords = ["cd2root", "Working Directory", "Project Root", "cd"]
classifiers = [
    #    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
dependencies = ["python-dotenv", "absl-py"]
dynamic = ["version", "readme"]

[project.urls]
Homepage = "https://github.com/garywei944/python-cd2root"
# Documentation = "https://readthedocs.org"
# Repository = "https://github.com/garywei944/python-cd2root"
Changelog = "https://github.com/garywei944/python-cd2root/blob/main/HISTORY.rst"
"Bug Tracker" = "https://github.com/garywei944/python-cd2root/issues"

[project.optional-dependencies]

[tool.setuptools.dynamic]
version = { attr = "cd2root.__version__" }
readme = { file = ["README.rst", "HISTORY.rst"] }
